+++ CHANGELOG.md
@@ -11,7 +11,9 @@ CHANGELOG for FlatCAM beta
 
 - Paint Tool - upgraded the UI and added the functionality that now adding a new tool is done by first searching the Tools DB for a suitable tool and if fails then it adds an default tool
 - Paint Tool - on start will attempt to search in the Tools DB for the default tools and if found will load them from the DB
-
+- NCC Tool - upgraded the UI and added the functionality that now adding a new tool is done by first searching the Tools DB for a suitable tool and if fails then it adds an default tool
+- NCC Tool - on start will attempt to search in the Tools DB for the default tools and if found will load them from the DB
+- fixes in NCC, Paint and Isolation Tool due of recent changes
 
 27.08.2020